Add snapcraft.yaml defining the GDC snap package and its components
The GDC package uses `classic` confinement to ensure that it can access
the resources of the host system it is installed on, e.g. libc and any
development libraries against which compiled code needs to link.

The `gdc` command is exposed together with the `gdmd` DMD-alike wrapper
script (downloaded in a separate part).

The build itself is managed via the `gcc` part: a `prepare` scriptlet
runs GDC's `setup-gcc.sh` script, and then we run the regular GCC build
via the `autotools` plugin, configured to enable D and LTO.